requirements.txt
# Core dependencies
flask>=3.0.0
pypdf>=4.0.0

# ML/NLP libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
llama-cpp-python>=0.2.0  # For running Mistral-7B locally

# Utilities
numpy>=1.24.0
scikit-learn>=1.3.0
huggingface_hub>=0.20.0  # For downloading models
